wealth_agent/requirements.txt
openai>=1.14.0
requests>=2.31.0
python-dotenv>=1.0.1
pandas>=2.2.1
beautifulsoup4>=4.12.3
streamlit>=1.32.0
gspread>=6.0.0
oauth2client>=4.1.3
git+https://github.com/openai/swarm.git
mcp>=1.0.0
playwright>=1.42.0
pyyaml>=6.0.1
streamlit-calendar>=0.1.0
